+++ lib/fishstick_dart.dart
@@ -14,5 +14,6 @@ export "commands/fortnite/account/login.dart";
 export "commands/fortnite/account/logout.dart";
 export "commands/fortnite/account/account.dart";
 export "commands/fortnite/vbucks/vbucks.dart";
+export "commands/fortnite/sac/sac.dart";
 
 final Client client = Client();