+++ CHANGELOG.md
@@ -56,3 +56,4 @@
 
 -   Auto sign out on invalid credentials.
 -   Abstract system job manager.
+-   Affiliate command.